pyproject.toml
[project]
name = "employee_verification_app"
version = "0.1.0"
description = "employee_verification_app using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.186.1,<1.0.0",
    "pyyaml>=6.0.1",
    "pypdf>=4.2.0",
    "pillow>=10.4.0",
    "setuptools>=69.0.0",
    "python-dotenv>=1.0.1",
    "pypdfium2>=4.30.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "langfuse>=3.10.1",
    "openinference-instrumentation-crewai>=0.1.16",
]

[project.scripts]
employee_verification_app = "employee_verification_app.main:run"
run_crew = "employee_verification_app.main:run"
train = "employee_verification_app.main:train"
replay = "employee_verification_app.main:replay"
test = "employee_verification_app.main:test"
serve_api = "employee_verification_app.server:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.uv]
constraint-dependencies = [
    "onnxruntime<1.23.0; platform_system == 'Darwin' and platform_machine == 'x86_64'",
    "pylance<0.38.0; platform_system == 'Darwin' and platform_machine == 'x86_64'",
]
